play a bit with test

+++ app.py
@@ -42,18 +42,19 @@ def after_request(response):
 
 class Test(Resource):
     def get(self):
-        '''fbc = FirebaseCon()
+        fbc = FirebaseCon()
         bucket = fbc.storage.bucket()
 
         # moving from temp to permanent path
         source = bucket.blob('source/source.jpg')
-        destination = bucket.blob('dest/dest.jpg')
-        destination.rewrite(source)
-        source.delete()
+        if source.exists():
+            destination = bucket.blob('dest/dest.jpg')
+            destination.rewrite(source)
+            source.delete()
         # could also use bucket.delete_blobs(['source', 'source'])
 
         # cleanup old temporary
-        oldTemp = bucket.list_blobs(prefix='2019-08-19/')
+        '''oldTemp = bucket.list_blobs(prefix='2019-08-19/')
         oldBlobs = list(oldTemp)
         print(len(oldBlobs))
         bucket.delete_blobs(oldBlobs)'''
